4.0 backport of rm_rf fixes

+++ conda/install.py
@@ -38,14 +38,18 @@ import subprocess
 import sys
 import tarfile
 import time
+import tempfile
 import traceback
 from os.path import (abspath, basename, dirname, isdir, isfile, islink, join)
 
 try:
     from conda.lock import Locked
+    from conda.config import pkgs_dirs
 except ImportError:
     # Make sure this still works as a standalone script for the Anaconda
     # installer.
+    pkgs_dirs = [sys.prefix]
+
     class Locked(object):
         def __init__(self, *args, **kwargs):
             pass
@@ -161,45 +165,40 @@ def rm_rf(path, max_retries=5, trash=True):
         # islink('/path/to/dead-link') is True.
         try:
             os.unlink(path)
+            return
         except (OSError, IOError):
             log.warn("Cannot remove, permission denied: {0}".format(path))
+            if trash and move_path_to_trash(path):
+                return
 
     elif isdir(path):
+
+        # On Windows, always move to trash first.
+        if trash and on_win and move_path_to_trash(path, preclean=False):
+            return
+
         for i in range(max_retries):
             try:
                 shutil.rmtree(path, ignore_errors=False, onerror=warn_failed_remove)
                 return
             except OSError as e:
+                if trash and move_path_to_trash(path):
+                    return
                 msg = "Unable to delete %s\n%s\n" % (path, e)
                 if on_win:
                     try:
                         shutil.rmtree(path, onerror=_remove_readonly)
                         return
-                    except OSError as e1:
-                        msg += "Retry with onerror failed (%s)\n" % e1
+                    except OSError as e2:
+                        raise
+                        msg += "Retry with onerror failed (%s)\n" % e2
 
-                    p = subprocess.Popen(['cmd', '/c', 'rd', '/s', '/q', path],
-                        stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-                    (stdout, stderr) = p.communicate()
-                    if p.returncode != 0:
-                        msg += '%s\n%s\n' % (stdout, stderr)
-                    else:
-                        if not isdir(path):
-                            return
+            log.debug(msg + "Retrying after %s seconds..." % i)
+            time.sleep(i)
 
-                    if trash:
-                        try:
-                            move_path_to_trash(path)
-                            if not isdir(path):
-                                return
-                        except OSError as e2:
-                            raise
-                            msg += "Retry with onerror failed (%s)\n" % e2
-
-                log.debug(msg + "Retrying after %s seconds..." % i)
-                time.sleep(i)
         # Final time. pass exceptions to caller.
         shutil.rmtree(path, ignore_errors=False, onerror=warn_failed_remove)
+
 
 def rm_empty_dir(path):
     """
@@ -543,9 +542,10 @@ def is_linked(prefix, dist):
 
 
 def delete_trash(prefix=None):
-    from conda.config import pkgs_dirs
     for pkg_dir in pkgs_dirs:
         trash_dir = join(pkg_dir, '.trash')
+        if not isdir(trash_dir):
+            continue
         try:
             log.debug("Trying to delete the trash dir %s" % trash_dir)
             rm_rf(trash_dir, max_retries=1, trash=False)
@@ -562,15 +562,14 @@ def move_to_trash(prefix, f, tempdir=None):
     """
     return move_path_to_trash(join(prefix, f))
 
-def move_path_to_trash(path):
+def move_path_to_trash(path, preclean=True):
     """
     Move a path to the trash
     """
     # Try deleting the trash every time we use it.
-    delete_trash()
-    import tempfile
+    if preclean:
+        delete_trash()
 
-    from conda.config import pkgs_dirs
     for pkg_dir in pkgs_dirs:
         trash_dir = join(pkg_dir, '.trash')
 
@@ -583,13 +582,15 @@ def move_path_to_trash(path):
         trash_file = tempfile.mktemp(dir=trash_dir)
 
         try:
-            shutil.move(path, trash_file)
+            os.rename(path, trash_file)
         except OSError as e:
             log.debug("Could not move %s to %s (%s)" % (path, trash_file, e))
         else:
+            log.debug("Moved to trash: %s" % (path,))
+            if not preclean:
+                rm_rf(trash_file, max_retries=1, trash=False)
             return True
 
-    log.debug("Could not move %s to trash" % path)
     return False
 
 # FIXME This should contain the implementation that loads meta, not is_linked()
@@ -623,17 +624,7 @@ def link(pkgs_dir, prefix, dist, linktype=LINK_HARD, index=None):
                 os.makedirs(dst_dir)
             if os.path.exists(dst):
                 log.warn("file already exists: %r" % dst)
-                try:
-                    os.unlink(dst)
-                except OSError:
-                    log.error('failed to unlink: %r' % dst)
-                    if on_win:
-                        try:
-                            move_path_to_trash(dst)
-                        except ImportError:
-                            # This shouldn't be an issue in the installer anyway
-                            pass
-
+                rm_rf(dst)
             lt = linktype
             if f in has_prefix_files or f in no_link or islink(src):
                 lt = LINK_COPY
@@ -705,17 +696,7 @@ def unlink(prefix, dist):
         for f in meta['files']:
             dst = join(prefix, f)
             dst_dirs1.add(dirname(dst))
-            try:
-                os.unlink(dst)
-            except OSError:  # file might not exist
-                log.debug("could not remove file: '%s'" % dst)
-                if on_win and os.path.exists(join(prefix, f)):
-                    try:
-                        log.debug("moving to trash")
-                        move_path_to_trash(dst)
-                    except ImportError:
-                        # This shouldn't be an issue in the installer anyway
-                        pass
+            rm_rf(dst)
 
         # remove the meta-file last
         os.unlink(meta_path)
@@ -805,6 +786,8 @@ def main():
 
     prefix = opts.prefix
     pkgs_dir = join(prefix, 'pkgs')
+    global pkgs_dirs
+    pkgs_dirs = [pkgs_dir]
     if opts.verbose:
         print("prefix: %r" % prefix)
